task 0 done 1

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	long unsigned num_node;
+	unsigned int num_node;
 
 	while (h)
 	{
